Module bee_storage_rocksdb::storage

Structs

Storage

Constants

CF_ADDRESS_TO_BALANCE

CF_ED25519_ADDRESS_TO_OUTPUT_ID

CF_INDEX_TO_MESSAGE_ID

CF_LEDGER_INDEX

CF_MESSAGE_ID_TO_MESSAGE

CF_MESSAGE_ID_TO_MESSAGE_ID

CF_MESSAGE_ID_TO_METADATA

CF_MILESTONE_INDEX_TO_MILESTONE

CF_MILESTONE_INDEX_TO_OUTPUT_DIFF

CF_OUTPUT_ID_TO_CONSUMED_OUTPUT

CF_OUTPUT_ID_TO_CREATED_OUTPUT

CF_OUTPUT_ID_UNSPENT

CF_SNAPSHOT_INFO

CF_SOLID_ENTRY_POINT_TO_MILESTONE_INDEX

CF_SYSTEM

Traits

Fetch
Fetch<K, V> trait extends the StorageBackend with fetch operation for the (key: K, value: V pair); therefore, it should be explicitly implemented for the corresponding StorageBackend.

Insert
Insert<K, V> trait extends the StorageBackend with insert operation for the (key: K, value: V) pair; therefore, it should be explicitly implemented for the corresponding StorageBackend.

StorageBackend
Trait to be implemented on a storage backend. Determines how to start and shutdown the backend.